+++ packages/components/src/components/ContentfulContainer/ContentfulContainerAsHyperlink.tsx
@@ -8,10 +8,10 @@ import type {
   StyleProps,
 } from '@contentful/experience-builder-core/types';
 
-import { EntityStore } from '@contentful/visual-sdk';
+import { EntityStore } from '@contentful/experience-builder-core';
 import { combineClasses } from '../../utils/combineClasses';
 
-export type ContentfulContainerProps<EditorMode = boolean> = EditorMode extends true
+export type ContentfulContainerAsHyperlinkProps<EditorMode = boolean> = EditorMode extends true
   ? {
       children?: React.ReactNode;
       className?: string;
@@ -37,7 +37,9 @@ export type ContentfulContainerProps<EditorMode = boolean> = EditorMode extends 
       children?: React.ReactNode;
     };
 
-export const ContentfulContainerAsHyperlink: React.FC<ContentfulContainerProps> = (props) => {
+export const ContentfulContainerAsHyperlink: React.FC<ContentfulContainerAsHyperlinkProps> = (
+  props
+) => {
   const { cfHyperlink, cfOpenInNewTab, editorMode, className, children } = props;
 
   let anchorTagProps = {};