+++ packages/core/src/entity/EditorEntityStore.ts
@@ -1,22 +1,7 @@
 import type { Asset, Entry } from 'contentful';
-
-import { EntityStore } from './EntityStore';
-
-export type RequestEntitiesMessage = {
-  entityIds: string[];
-  entityType: 'Asset' | 'Entry';
-  locale: string;
-};
-
-export type RequestedEntitiesMessage = {
-  entities: Array<Entry | Asset>;
-  missingEntityIds?: string[];
-};
-
-export enum PostMessageMethods {
-  REQUEST_ENTITIES = 'REQUEST_ENTITIES',
-  REQUESTED_ENTITIES = 'REQUESTED_ENTITIES',
-}
+import { EntityStoreBase } from './EntityStoreBase';
+import { PostMessageMethods } from '../constants';
+import { RequestEntitiesMessage, RequestedEntitiesMessage } from '../types';
 
 type SendMessage = (
   method: PostMessageMethods.REQUEST_ENTITIES,
@@ -31,7 +16,7 @@ type Subscribe = (
  * EntityStore which resolves entries and assets from the editor
  * over the sendMessage and subscribe functions.
  */
-export class EditorEntityStore extends EntityStore {
+export class EditorEntityStore extends EntityStoreBase {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   private requestCache = new Map<string, Promise<any>>();
   private sendMessage: SendMessage;
@@ -69,17 +54,32 @@ export class EditorEntityStore extends EntityStore {
     return id.length === 1 ? id[0] : id.join(this.cacheIdSeperator);
   }
 
-  private async fetchEntity(type: 'Asset', ids: string[]): Promise<Array<Asset>>;
-  private async fetchEntity(type: 'Entry', ids: string[]): Promise<Array<Entry>>;
+  private async fetchEntity(
+    type: 'Asset',
+    ids: string[],
+    skipCache: boolean
+  ): Promise<Array<Asset>>;
+  private async fetchEntity(
+    type: 'Entry',
+    ids: string[],
+    skipCache: boolean
+  ): Promise<Array<Entry>>;
   private async fetchEntity(
     type: 'Asset' | 'Entry',
-    ids: string[]
+    ids: string[],
+    skipCache: boolean = false
   ): Promise<Array<Entry> | Array<Asset>> {
-    const { missing, resolved } = this.getEntitiesFromMap(type, ids);
+    let missing: string[];
+    if (!skipCache) {
+      const { missing: missingFromCache, resolved } = this.getEntitiesFromMap(type, ids);
+      if (missingFromCache.length === 0) {
+        // everything is already in cache
+        return resolved as Array<Entry> | Array<Asset>;
+      }
 
-    if (missing.length === 0) {
-      // everything is already in cache
-      return resolved as Array<Entry> | Array<Asset>;
+      missing = missingFromCache;
+    } else {
+      missing = [...ids];
     }
 
     const cacheId = this.getCacheId(missing);
@@ -145,9 +145,9 @@ export class EditorEntityStore extends EntityStore {
     return this.getEntitiesFromMap(type, ids).resolved as Array<Entry> | Array<Asset>;
   }
 
-  public async fetchAsset(id: string): Promise<Asset | undefined> {
+  public async fetchAsset(id: string, skipCache: boolean = false): Promise<Asset | undefined> {
     try {
-      return (await this.fetchAssets([id]))[0];
+      return (await this.fetchAssets([id], skipCache))[0];
     } catch (err) {
       // TODO: move to debug utils once it is extracted
       console.warn(`Failed to request asset ${id}`);
@@ -155,13 +155,13 @@ export class EditorEntityStore extends EntityStore {
     }
   }
 
-  public fetchAssets(ids: string[]): Promise<Asset[]> {
-    return this.fetchEntity('Asset', ids);
+  public fetchAssets(ids: string[], skipCache: boolean = false): Promise<Asset[]> {
+    return this.fetchEntity('Asset', ids, skipCache);
   }
 
-  public async fetchEntry(id: string): Promise<Entry | undefined> {
+  public async fetchEntry(id: string, skipCache: boolean = false): Promise<Entry | undefined> {
     try {
-      return (await this.fetchEntries([id]))[0];
+      return (await this.fetchEntries([id], skipCache))[0];
     } catch (err) {
       // TODO: move to debug utils once it is extracted
       console.warn(`Failed to request entry ${id}`, err);
@@ -169,7 +169,7 @@ export class EditorEntityStore extends EntityStore {
     }
   }
 
-  public fetchEntries(ids: string[]): Promise<Entry[]> {
-    return this.fetchEntity('Entry', ids);
+  public fetchEntries(ids: string[], skipCache: boolean = false): Promise<Entry[]> {
+    return this.fetchEntity('Entry', ids, skipCache);
   }
 }