+++ packages/core/src/entity/EditorModeEntityStore.ts
@@ -1,6 +1,7 @@
-import { EditorEntityStore, RequestedEntitiesMessage } from '@contentful/visual-sdk';
 import type { Asset, AssetFile, Entry, UnresolvedLink } from 'contentful';
 import { sendMessage } from '../communication/sendMessage';
+import { EditorEntityStore } from './EditorEntityStore';
+import { RequestedEntitiesMessage } from '../types';
 
 // The default of 3s in the EditorEntityStore is sometimes timing out and
 // leads to not rendering bound content and assemblies.
@@ -9,7 +10,7 @@ const REQUEST_TIMEOUT = 10000;
 export class EditorModeEntityStore extends EditorEntityStore {
   public locale: string;
 
-  constructor({ entities, locale }: { entities: Array<Entry | Asset>; locale: string }) {
+  constructor({ entities, locale }: { entities: Array<Asset | Entry>; locale: string }) {
     console.debug(
       `[exp-builder.sdk] Initializing editor entity store with ${entities.length} entities for locale ${locale}.`,
       { entities }
@@ -55,12 +56,17 @@ export class EditorModeEntityStore extends EditorEntityStore {
   async fetchEntities({
     missingEntryIds,
     missingAssetIds,
+    skipCache = false,
   }: {
     missingEntryIds: string[];
     missingAssetIds: string[];
+    skipCache?: boolean;
   }) {
     // Entries and assets will be stored in entryMap and assetMap
-    await Promise.all([this.fetchEntries(missingEntryIds), this.fetchAssets(missingAssetIds)]);
+    await Promise.all([
+      this.fetchEntries(missingEntryIds, skipCache),
+      this.fetchAssets(missingAssetIds, skipCache),
+    ]);
   }
 
   getMissingEntityIds(entityLinks: UnresolvedLink<'Entry' | 'Asset'>[]) {