+++ packages/core/src/entity/index.ts
@@ -1,2 +1,3 @@
 export * from './EditorModeEntityStore';
 export * from './EntityStore';
+export * from './EntityStoreBase';