+++ packages/visual-editor/src/components/Dropzone/Dropzone.tsx
@@ -1,4 +1,4 @@
-import React, { ElementType, useCallback, useEffect, useMemo } from 'react';
+import React, { ElementType, useCallback, useEffect } from 'react';
 import { Droppable } from '@hello-pangea/dnd';
 import type {
   CompositionComponentNode,
@@ -12,7 +12,7 @@ import { usePlaceholderStyleStore } from '@/store/placeholderStyle';
 import styles from './styles.module.css';
 import classNames from 'classnames';
 import { ROOT_ID } from '@/types/constants';
-import { EmptyEditorContainer } from '@components/EmptyContainer/EmptyContainer';
+import { EmptyContainer } from '@components/EmptyContainer/EmptyContainer';
 import { getZoneParents } from '@/utils/zone';
 import { useZoneStore } from '@/store/zone';
 import { useDropzoneDirection } from '@/hooks/useDropzoneDirection';
@@ -37,7 +37,6 @@ function isDropEnabled(
   hoveringOverSection: boolean,
   draggingRootZone: boolean,
   isRootZone: boolean,
-  draggingOverArea: boolean,
   isAssembly: boolean
 ) {
   if (isAssembly) {
@@ -60,7 +59,7 @@ function isDropEnabled(
     return isRootZone;
   }
 
-  return draggingOverArea;
+  return userIsDragging;
 }
 
 export function Dropzone({
@@ -100,14 +99,6 @@ export function Dropzone({
     addSectionWithZone(sectionId);
   }, [sectionId, addSectionWithZone]);
 
-  const draggingOverArea = useMemo(() => {
-    if (!userIsDragging) {
-      return false;
-    }
-
-    return draggingParentIds[0] === zoneId;
-  }, [userIsDragging, draggingParentIds, zoneId]);
-
   const isAssembly =
     DESIGN_COMPONENT_NODE_TYPES.includes(node?.type || '') ||
     ASSEMBLY_NODE_TYPES.includes(node?.type || '');
@@ -127,7 +118,6 @@ export function Dropzone({
     hoveringOverSection,
     draggingRootZone,
     isRootZone,
-    draggingOverArea,
     isAssembly
   );
 
@@ -151,8 +141,17 @@ export function Dropzone({
     return null;
   }
 
+  // Don't trigger the dropzone when it's the root because then the only hit boxes that show up will be root level zones
+  // Exception 1: If it comes from the component list (because we want the component list components to work for all zones
+  // Exception 2: If it's a child of a root level zone (because we want to be able to re-order root level containers)
+  const isNotDroppable =
+    zoneId === ROOT_ID && draggedSourceId !== 'component-list' && draggingParentIds.length !== 0;
+
   return (
-    <Droppable droppableId={droppableId} direction={direction} isDropDisabled={!dropEnabled}>
+    <Droppable
+      droppableId={droppableId}
+      direction={direction}
+      isDropDisabled={!dropEnabled || isNotDroppable}>
       {(provided, snapshot) => {
         return (
           <WrapperComponent
@@ -180,7 +179,7 @@ export function Dropzone({
             }}
             {...rest}>
             {isEmptyCanvas ? (
-              <EmptyEditorContainer isDragging={isRootZone && userIsDragging} />
+              <EmptyContainer isDragging={isRootZone && userIsDragging} />
             ) : (
               content.map((item, i) => {
                 const componentId = item.data.id;