+++ packages/storybook-addon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-alpha.17](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.16...@contentful/experience-builder-storybook-addon@0.0.2-alpha.17) (2024-01-23)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
+## [0.0.2-alpha.16](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.15...@contentful/experience-builder-storybook-addon@0.0.2-alpha.16) (2024-01-17)
+
+**Note:** Version bump only for package @contentful/experience-builder-storybook-addon
+
 ## [0.0.2-alpha.15](https://github.com/contentful/experience-builder-storybook/compare/@contentful/experience-builder-storybook-addon@0.0.2-alpha.14...@contentful/experience-builder-storybook-addon@0.0.2-alpha.15) (2024-01-16)
 
 **Note:** Version bump only for package @contentful/experience-builder-storybook-addon